requirements.txt
flask
boto3
psycopg2-binary
joblib
scikit-learn
gunicorn
